Move commandtype to command folder
